+++ cli/cluster/cluster.go
@@ -7,8 +7,8 @@ import (
 	"time"
 
 	"github.com/tarantool/go-tarantool"
-	"github.com/tarantool/tt/cli/connect"
 	libcluster "github.com/tarantool/tt/lib/cluster"
+	"github.com/tarantool/tt/lib/connect"
 )
 
 const (
